+++ language-server/Shared.AnZwDev.ALTools.Server/Contracts/CollectWorkspaceCommandCodeActionsResponse.cs
@@ -1,0 +1,22 @@
+﻿using AnZwDev.ALTools.WorkspaceCommands;
+using Newtonsoft.Json;
+using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace AnZwDev.ALTools.Server.Contracts
+{
+    public class CollectWorkspaceCommandCodeActionsResponse
+    {
+
+        [JsonProperty("codeActions", NullValueHandling = NullValueHandling.Ignore)]
+        public List<WorkspaceCommandCodeAction> CodeActions { get; set; }
+
+        [JsonProperty("error")]
+        public bool Error { get; set; }
+
+        [JsonProperty("errorMessage", NullValueHandling = NullValueHandling.Ignore)]
+        public string ErrorMessage { get; set; }
+
+    }
+}