+++ vscode-extension/src/alsyntaxmodifiers/addMissingCaseLinesModifier.ts
@@ -1,0 +1,10 @@
+import { DevToolsExtensionContext } from "../devToolsExtensionContext";
+import { WorkspaceCommandSyntaxModifier } from "./workspaceCommandSyntaxModifier";
+
+export class AddMissingCaseLinesModifier extends WorkspaceCommandSyntaxModifier {
+
+    constructor(context: DevToolsExtensionContext) {
+        super(context, "Add Missing Case Lines", "addMissingCaseLines");
+    }
+
+}