+++ vscode-extension/src/langserver/toolsWorkspaceCommandCodeAction.ts
@@ -1,0 +1,7 @@
+import { TextRange } from "../symbollibraries/textRange";
+
+export interface ToolsWorkspaceCommandCodeAction {
+    commandName: string | undefined;
+    description: string | undefined;
+    range: TextRange | undefined;
+}